Changes to cover images are now cancelled when you click "Cancel". This applies to both changing the cover image and setting it to "No Cover". In both cases, if you click "Cancel" now, the changes are reverted to the previous saved state of the book. Fixes bug #16707
git-svn-id: http://alexandria.rubyforge.org/svn/trunk/alexandria@1053 9992d806-23ff-4cc5-b4c6-abbc17c6a14e

+++ lib/alexandria/ui/dialogs/new_book_dialog_manual.rb
@@ -132,7 +132,7 @@ module Alexandria
 
             @library << book
             @library.save(book)
-            if File.exists?(TMP_COVER_FILE)
+            if (File.exists?(TMP_COVER_FILE) and (not @delete_cover_file))
               FileUtils.cp(TMP_COVER_FILE, @library.cover(book))
             end
